[luci] Rename FuseBatchNormWithTConv files
This will rename FuseBatchNormWithTConv to FuseBatchNormWithTConvPass to
follow like other file naming convention.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
